serving/requirements.txt
fastapi
uvicorn[standard]
pandas
pyarrow
apscheduler
scipy
python-dotenv
pydantic
requests
numpy
